main-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy>=2.0.28
pymongo==4.6.0
motor==3.3.1
pydantic[email]>=2.5.2
python-dotenv==1.0.0
psycopg2-binary
alembic==1.12.1
asyncpg>=0.29.0
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic-settings
tenacity